backend/requirements.txt
langgraph>=0.2.0
langchain
langchain-openai
fastapi
uvicorn
python-dotenv
pydantic
